+++ packages/database/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @allends/graphene-database
 
+## 0.1.5
+
+### Patch Changes
+
+- Update build
+
+## 0.1.4
+
+### Patch Changes
+
+- Add binary back
+
 ## 0.1.3
 
 ### Patch Changes